requirements.txt
Pillow
exifread
PyPDF2
python-docx